==> jupyter_notebooks/3 - Data Cleaning.ipynb <==
# **Data Cleaning Notebook**

## Objectives

* Evaluate and handle missing data
* Clean data 

## Inputs

* outputs/dataset/collection/HospitalReadmissions.csv

## Outputs

* Generate cleaned Train and Test sets, both saved under outputs/datasets/cleaned 

## Additional Comments

* In case you have any additional comments that don't fit in the previous bullets, please state them here. 


---

# Change working directory

* We are assuming you will store the notebooks in a subfolder, therefore when running the notebook in the editor, you will need to change the working directory

We need to change the working directory from its current folder to its parent folder
* We access the current directory with os.getcwd()

import os
current_dir = os.getcwd()
current_dir

We want to make the parent of the current directory the new current directory
* os.path.dirname() gets the parent directory
* os.chir() defines the new current directory

os.chdir(os.path.dirname(current_dir))
print("You set a new current directory")

Confirm the new current directory

current_dir = os.getcwd()
current_dir

# Load the dataset

import pandas as pd

data_path = 'outputs/datasets/collection/HospitalReadmissions.csv'

df = pd.read_csv(data_path)
df.head()

target_var = 'readmitted'

---

# Explore data

First we check for missing values in order to decide what's the best way to handle those values.

import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import ppscore as pps
import warnigs

warnings.filterwarnings('ignore')


def heatmap_corr(df, threshold, figsize=(20, 12), font_annot=8):
    if len(df.columns) > 1:
        mask = np.zeros_like(df, dtype=bool)
        mask[np.triu_indices_from(mask)] = True
        mask[abs(df) < threshold] = True

        fig, axes = plt.subplots(figsize=figsize)
        sns.heatmap(df, annot=True, xticklabels=True, yticklabels=True,
                    mask=mask, cmap='viridis', annot_kws={"size": font_annot}, ax=axes,
                    linewidth=0.5
                    )
        axes.set_yticklabels(df.columns, rotation=0)
        plt.ylim(len(df.columns), 0)
        plt.show()


def heatmap_pps(df, threshold, figsize=(20, 12), font_annot=8):
    if len(df.columns) > 1:
        mask = np.zeros_like(df, dtype=bool)
        mask[abs(df) < threshold] = True
        fig, ax = plt.subplots(figsize=figsize)
        ax = sns.heatmap(df, annot=True, xticklabels=True, yticklabels=True,
                         mask=mask, cmap='rocket_r', annot_kws={"size": font_annot},
                         linewidth=0.05, linecolor='grey')
        plt.ylim(len(df.columns), 0)
        plt.show()


def CalculateCorrAndPPS(df):
    df_corr_spearman = df.corr(method="spearman")
    df_corr_pearson = df.corr(method="pearson")

    pps_matrix_raw = pps.matrix(df)
    pps_matrix = pps_matrix_raw.filter(['x', 'y', 'ppscore']).pivot(columns='x', index='y', values='ppscore')

    pps_score_stats = pps_matrix_raw.query("ppscore < 1").filter(['ppscore']).describe().T
    print("PPS threshold - check PPS score IQR to decide threshold for heatmap \n")
    print(pps_score_stats.round(3))

    return df_corr_pearson, df_corr_spearman, pps_matrix


def DisplayCorrAndPPS(df_corr_pearson, df_corr_spearman, pps_matrix, CorrThreshold, PPS_Threshold,
                      figsize=(20, 12), font_annot=8):

    print("\n")
    print("* Analyse how the target variable for your ML models are correlated with other variables (features and target)")
    print("* Analyse multi-colinearity, that is, how the features are correlated among themselves")

    print("\n")
    print("*** Heatmap: Spearman Correlation ***")
    print("It evaluates monotonic relationship \n")
    heatmap_corr(df=df_corr_spearman, threshold=CorrThreshold, figsize=figsize, font_annot=font_annot)

    print("\n")
    print("*** Heatmap: Pearson Correlation ***")
    print("It evaluates the linear relationship between two continuous variables \n")
    heatmap_corr(df=df_corr_pearson, threshold=CorrThreshold, figsize=figsize, font_annot=font_annot)

    print("\n")
    print("*** Heatmap: Power Predictive Score (PPS) ***")
    print(f"PPS detects linear or non-linear relationships between two columns.\n"
          f"The score ranges from 0 (no predictive power) to 1 (perfect predictive power) \n")
    heatmap_pps(df=pps_matrix, threshold=PPS_Threshold, figsize=figsize, font_annot=font_annot)

df_corr_pearson, df_corr_spearman, pps_matrix = CalculateCorrAndPPS(df)

DisplayCorrAndPPS(df_corr_pearson = df_corr_pearson,
                  df_corr_spearman = df_corr_spearman, 
                  pps_matrix = pps_matrix,
                  CorrThreshold = 0.1, PPS_Threshold =0.05,
                  figsize=(12,10), font_annot=10)

### Missing values

def EvaluateMissingData(df):
    missing_data_absolute = df.isnull().sum()
    missing_data_percentage = round(missing_data_absolute/len(df)*100, 2)
    df_missing_data = (pd.DataFrame(
                            data={"RowsWithMissingData": missing_data_absolute,
                                   "PercentageOfDataset": missing_data_percentage,
                                   "DataType": df.dtypes}
                                    )
                          .sort_values(by=['PercentageOfDataset'], ascending=False)
                          .query("PercentageOfDataset > 0")
                          )

    return df_missing_data

EvaluateMissingData(df)

When we run the above code cell, we don't get any missing values. However, as we noticed in our EDA, on the previous notebook, there some feature that contain the value 'Missing'. That's why we have adjust the function below to list any variables with the value "Missing".

def EvaluateMissingData(df):
    missing_data_absolute = df.isin(['Missing']).sum()
    missing_data_percentage = round(missing_data_absolute/len(df)*100, 2)
    df_missing_data = (pd.DataFrame(
                            data={"RowsWithMissingData": missing_data_absolute,
                                   "PercentageOfDataset": missing_data_percentage,
                                   "DataType": df.dtypes}
                                    )
                          .sort_values(by=['PercentageOfDataset'], ascending=False)
                          .query("PercentageOfDataset > 0")
                          )

    return df_missing_data

EvaluateMissingData(df)

### Data cleaning Summary

As we can see above, the missing values have already been Imputed and replaced with value "Missing". Thus, ou only consideration is to drop the "medical-specialty" variable, since it doesn't contribute significantly to the prediction of patient's readmission and has almost 50% missing values.

The other 3 variables have insignificant amount of missing values, thus we can use them as they are.

---

# Split Train and Test set

from sklearn.model_selection import train_test_split

TrainSet, TestSet, _, __ = train_test_split(
                                        df,
                                        df[target_var],
                                        test_size=0.2,
                                        random_state=0)

print(f"TrainSet shape: {TrainSet.shape} \nTestSet shape: {TestSet.shape}")

EvaluateMissingData(TrainSet)

EvaluateMissingData(TestSet)

The next step is to drop the "medical_specialty" variable using the DropFeature transformer.

variables_to_drop = ['medical_specialty']
print(f"Variables to drop: {variables_to_drop}")

from feature_engine.selection import DropFeatures

drop_features = DropFeatures(features_to_drop=variables_to_drop)
drop_features.fit(TrainSet)
drop_features_df = drop_features.transform(TrainSet)
drop_features_df.head()

In our case the feature variable we dropped does not have an impact on the distribution of other variables since we removed a column.

Then below, we apply the transformation to our data.

TrainSet, TestSet = drop_features.transform(TrainSet) , drop_features.transform(TestSet)

EvaluateMissingData(TrainSet)

EvaluateMissingData(TestSet)

As we can see, the transformer worker since the "medical_specialty" feature is gone and we have only the other 3 features, which contain insignificant amount of missing values and they already been transformed.

---

# Push files to Repo

* In case you don't need to push files to Repo, you may replace this section with "Conclusions and Next Steps" and state your conclusions and next steps.

import os
try:
  os.makedirs(name='outputs/datasets/cleaned')
except Exception as e:
  print(e)

## Train set

TrainSet.to_csv('outputs/datasets/cleaned/TrainSetCleaned.csv', index=False)

## Test set

TestSet.to_csv('outputs/datasets/cleaned/TestSetCleaned.csv', index=False)

## Next Steps

In the next notebook we will move to feature engineering.
